rename gcs placement group mgr
