+++ src/ray/gcs/gcs_server/BUILD.bazel
@@ -210,7 +210,7 @@ ray_cc_library(
         "gcs_actor_scheduler.cc",
         "gcs_autoscaler_state_manager.cc",
         "gcs_node_manager.cc",
-        "gcs_placement_group_manager.cc",
+        "gcs_placement_group_mgr.cc",
         "gcs_placement_group_scheduler.cc",
         "gcs_resource_manager.cc",
         "gcs_server.cc",
@@ -220,7 +220,7 @@ ray_cc_library(
         "gcs_actor_scheduler.h",
         "gcs_autoscaler_state_manager.h",
         "gcs_node_manager.h",
-        "gcs_placement_group_manager.h",
+        "gcs_placement_group_mgr.h",
         "gcs_placement_group_scheduler.h",
         "gcs_resource_manager.h",
         "gcs_server.h",
@@ -247,10 +247,10 @@ ray_cc_library(
         "//:gcs_service_rpc",
         "//:node_manager_rpc",
         "//:observable_store_client",
-        "//src/ray/pubsub:pubsub_lib",
         "//:raylet_client_lib",
         "//:worker_rpc",
         "//src/ray/protobuf:agent_manager_cc_proto",
+        "//src/ray/pubsub:pubsub_lib",
         "//src/ray/raylet/scheduling:scheduler",
         "//src/ray/util:counter_map",
         "//src/ray/util:thread_checker",
