+++ src/ray/gcs/gcs_server/gcs_autoscaler_state_manager.cc
@@ -21,7 +21,7 @@
 
 #include "ray/gcs/gcs_server/gcs_actor_manager.h"
 #include "ray/gcs/gcs_server/gcs_node_manager.h"
-#include "ray/gcs/gcs_server/gcs_placement_group_manager.h"
+#include "ray/gcs/gcs_server/gcs_placement_group_mgr.h"
 #include "ray/gcs/pb_util.h"
 
 namespace ray {
