+++ src/ray/gcs/gcs_server/gcs_placement_group_scheduler.cc
@@ -21,7 +21,7 @@
 #include <vector>
 
 #include "ray/common/asio/asio_util.h"
-#include "ray/gcs/gcs_server/gcs_placement_group_manager.h"
+#include "ray/gcs/gcs_server/gcs_placement_group_mgr.h"
 #include "src/ray/protobuf/gcs.pb.h"
 
 namespace ray {
