+++ src/ray/gcs/gcs_server/test/gcs_placement_group_mgr_test.cc
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// clang-format off
-#include "ray/gcs/gcs_server/gcs_placement_group_manager.h"
-// clang-format on
+#include "ray/gcs/gcs_server/gcs_placement_group_mgr.h"
 
 #include <memory>
 #include <string>
